変数名などリファクタリング

+++ mysite/account/views.py
@@ -4,53 +4,17 @@ from . import models
 from . import forms
 import logging 
 from .utils import is_login
-# Create your views here.
 
-logger = logging.getLogger('login') # loggerを指定
-
-def index(request): 
-    return redirect('shopping:index')
-
-def login(request): 
-    if request.session.get('is_login', None): 
-        return redirect('shopping:index')
-    if request.method == 'POST':
-        login_form = forms.UserForm(request.POST) #ログインForm生成 
-        message = '入力した内容を再度確認してください' 
-        if login_form.is_valid(): 
-            user_id = login_form.cleaned_data.get('id') 
-            password = login_form.cleaned_data.get('password') 
-            try: 
-                user = models.User.objects.get(user_id=user_id) 
-            except: 
-                message = 'ユーザが存在しません' 
-                return render(request, 'account/login.html', locals()) #この環境内で定義されている変数を辞書として取得できる 
-            if user.password == password: 
-                request.session['is_login'] = True #session書き込み 
-                request.session['user_id'] = user.user_id 
-                request.session['user_name'] = user.name 
-                logger.info('url:%s method:%s user:%s login'% (request.path, request.method, user.user_id)) #log 
-                return redirect('shopping:index')
-            else: 
-                message = 'パスワードが正しくありません。'
-                return render(request, 'account/login.html', locals())
-        else: 
-            return render(request, 'account/login.html', locals())
-    login_form = forms.UserForm()
-    return render(request, 'account/login.html', locals())
-
-@is_login
-def logout(request):
-    request.session.flush() #session削除 
-    return redirect('account:login')
+logger = logging.getLogger('login')
 
 def register_user(request):
-    if request.session.get('is_login', None): #ログインしていない状態確認
+    if request.session.get('is_login', None):
         return redirect('shopping:index')
+    
     if request.method == 'POST':
         register_form = forms.RegisterForm(request.POST) 
         message = "入力した内容を再度確認してください"
-        if register_form.is_valid(): #form検査
+        if register_form.is_valid():
             user_id = register_form.cleaned_data.get('id') 
             password1 = register_form.cleaned_data.get('password1')
             password2 = register_form.cleaned_data.get('password2')
@@ -69,10 +33,11 @@ def register_user(request):
             return render(request, 'account/registerUser.html', locals()) 
     register_form = forms.RegisterForm() 
     return render(request, 'account/registerUser.html', locals()) 
-    
+
 def register_user_commit(request):
     if request.session.get('is_login', None):
         return redirect('shopping:index')
+    
     if request.method == 'POST':
         register_form = forms.RegisterForm(request.POST) 
         if register_form.is_valid():
@@ -93,22 +58,57 @@ def register_user_commit(request):
             return render(request, 'account/registerUserCommit.html')
         else:
             return render(request, 'account/registerUser.html', locals()) 
-    
     register_form = forms.RegisterForm()
     return render(request, 'account/registerUser.html', locals())
+
+def login(request): 
+    if request.session.get('is_login', None): 
+        return redirect('shopping:index')
+    if request.method == 'POST':
+        login_form = forms.UserForm(request.POST)
+        message = '入力した内容を再度確認してください' 
+        if login_form.is_valid(): 
+            user_id = login_form.cleaned_data.get('id') 
+            password = login_form.cleaned_data.get('password') 
+            try: 
+                user = models.User.objects.get(user_id=user_id) 
+            except: 
+                message = 'ユーザが存在しません' 
+                return render(request, 'account/login.html', locals())
+            if user.password == password: 
+                request.session['is_login'] = True 
+                request.session['user_id'] = user.user_id 
+                request.session['user_name'] = user.name 
+                logger.info('url:%s method:%s user:%s login'% (request.path, request.method, user.user_id))
+                return redirect('shopping:index')
+            else: 
+                message = 'パスワードが正しくありません。'
+                return render(request, 'account/login.html', locals())
+        else: 
+            return render(request, 'account/login.html', locals())
+    login_form = forms.UserForm()
+    return render(request, 'account/login.html', locals())
+
+@is_login
+def logout(request):
+    request.session.flush()
+    return redirect('account:login')
 
 @is_login
 def user_info(request):
     user_id = request.session['user_id']
     user = models.User.objects.get(user_id=user_id)
-    id = user_id
-    name = user.name 
-    address = user.address
-    register_form = forms.RegisterForm()
     return render(request, 'account/userInfo.html', locals())
 
 @is_login
 def update_user(request): 
+    user_id = request.session['user_id']
+    user = models.User.objects.get(user_id=user_id)
+    register_form = forms.RegisterForm(initial={
+        'id': user.user_id,
+        'name': user.name,
+        'address': user.address
+        })
     if request.method == 'POST':
         register_form = forms.RegisterForm(request.POST)
         message = "入力した内容を再度確認してください" 
@@ -124,15 +124,9 @@ def update_user(request):
             else: 
                 return render(request, 'account/updateUserConfirm.html', locals()) 
         else: 
-            return render(request, 'account/updateUser.html', locals()) 
-    user_id = request.session['user_id']
-    user = models.User.objects.get(user_id=user_id) 
-    initial = True 
-    id = user_id 
-    name = user.name 
-    address = user.address 
-    register_form = forms.RegisterForm() 
-    return render(request, 'account/updateUser.html', locals()) 
+            message = '入力された内容に誤りがあります'
+            return render(request, 'account/updateUser.html', locals())
+    return render(request, 'account/updateUser.html', locals())
             
 @is_login
 def update_user_commit(request): 
@@ -153,17 +147,16 @@ def update_user_commit(request):
             return render(request, 'account/updateUserCommit.html', locals()) 
         else: 
             return render(request, 'account/updateUser.html', locals()) 
-    return render(request, 'account/main.html') 
+    return render(request, 'account/updateUser.html', locals())
             
 @is_login
 def withdraw(request):
+    if request.method == 'POST':
+        user_id = request.session['user_id']
+        user = models.User.objects.get(user_id=user_id)
+        user_name_cashe = models.User.objects.get(user_id=user_id).name
+        user.delete()
+        request.session.flush() 
+        return render(request, 'account/withdrawCommit.html', locals())
     return render(request, 'account/withdrawConfirm.html')
     
-@is_login
-def withdraw_commit(request):
-    user_id = request.session['user_id']
-    name = request.session['user_name']
-    user = models.User.objects.get(user_id=user_id)
-    user.delete()
-    request.session.flush() 
-    return render(request, 'account/withdrawCommit.html', locals())
